test(business-simulations): add 68 unit tests for BusinessSimulations BC (Issue #3727) (#4039)
- CostScenarioTests: 17 tests for domain entity (Create factory, validation, edge cases)
- AgentCostValidatorTests: 31 tests for all 4 FluentValidation validators
- LlmCostCalculatorTests: 20 tests for cost calculation service (pricing, precision, edge cases)
- Fix enterprise-navigation test: Business section has 4 tabs (was 3)
- Fix Sonar S1905: remove redundant long casts in EstimateResourceForecastQueryHandler

Refs: #3727

+++ apps/api/src/Api/BoundedContexts/BusinessSimulations/Application/Handlers/EstimateResourceForecastQueryHandler.cs
@@ -59,13 +59,13 @@ internal sealed class EstimateResourceForecastQueryHandler
             var projectedDbGb = query.CurrentDbSizeGb + (additionalDbMb / 1024m);
 
             var projectedDailyTokens = query.CurrentDailyTokens +
-                ((long)newUsers * query.TokensPerUserPerDay);
+                (newUsers * query.TokensPerUserPerDay);
 
             var additionalCacheMb = newUsers * query.CachePerUserMb;
             var projectedCacheMb = query.CurrentCacheMb + additionalCacheMb;
 
             var projectedVectors = query.CurrentVectorEntries +
-                ((long)newUsers * query.VectorsPerUser);
+                (newUsers * query.VectorsPerUser);
 
             var monthlyCost = CalculateMonthlyCost(
                 projectedDbGb, projectedDailyTokens, projectedCacheMb, projectedVectors);
